Organize hooks in folders
