+++ packages/circuit-ui/hooks/useClickOutside/useClickOutside.ts
