+++ packages/circuit-ui/hooks/useFocusList/useFocusList.spec.tsx
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-import { render, act, userEvent } from '../util/test-utils';
+import { render, act, userEvent } from '../../util/test-utils';
 
 import { useFocusList } from './useFocusList';
